Shuts down properly
Context disposal was hanging because there are still open sockets in the logger.
Change the Publish in the logger to open own socket and dispose, that way we know there are no open sockets when we dispose of context.

To avoid exception when the logger is called, don't try to open a socket unless we are running.